Update to newest version of the simulator

+++ structure_optimizer/src/main/java/u_szeged/inf/fog/structure_optimizer/models/SimulationResult.java
@@ -23,7 +23,7 @@ public class SimulationResult {
     @NotNull
     private double totalEnergyConsumption;
     @NotNull
-    private long executionTime;
+    private double executionTime;
 
     @NotNull
     private int totalTasks;
